Overall: Fixed warnings

+++ Artificial/ConsoleWindow.cpp
@@ -40,8 +40,8 @@ namespace GUI
 	{
 		CONSOLE_SCREEN_BUFFER_INFO info;
 		GetConsoleScreenBufferInfo(m_console_handler, &info);
-		Size size {info.srWindow.Bottom - info.srWindow.Top + 1,
-				   info.srWindow.Right - info.srWindow.Left + 1};
+		Size size {uint16_t(info.srWindow.Bottom - info.srWindow.Top + 1),
+				   uint16_t(info.srWindow.Right - info.srWindow.Left + 1)};
 		return size;
 	}
